# api-doc.yaml
swagger: '2.0'
info:
  description: 'Webservice documentation for the LCA Collaboration Server'
  version: '1.1.1'
  title: 'LCA Collaboration Server Webservice'
  contact:
    email: 'greve@greendelta.com'
host: 'data.openlca.org'
schemes:
- https
- http
definitions:
...
  AggregationResultEntry:
    type: object
    properties: 
      key:
        description: 'The key of this aggregation result entry'
        type: string
        example: 'ACTOR'
      count:
        description: 'The amount of entries found for the given key'
        type: integer
        format: int64
        example: '12'
      subEntries:
        description: 'Children elements in case of hierarchical entries'
        type: array
        items: 
          $ref: '#/definitions/AggregationResultEntry'
    required:
     - key
     - count
  AggregationResult:
    type: object
    properties: 
      name:
        description: 'The name of this aggregation result'
        type: string
        example: 'repositoryId'
      type:
        description: 'The type of aggregation'
        type: string
        example: 'TERM'
      totalCount:
        description: 'The amount of entries found for this aggregation'
        type: integer
        format: int64
        example: '13'
      entries:
        description: 'A list of entries into which the aggregation splits (each possible result value is a result entry)'
        type: array
        items: 
          $ref: '#/definitions/AggregationResultEntry'
    required:
      - name
      - type
      - totalCount
  ResultInfo:
    type: object
    properties: 
      pageSize:
        description: 'The size of each page of results'
        type: integer
        format: int32
        example: '10'
      count:
        description: 'The amount of actual results returned'
        type: integer
        format: int64
        example: '8'
      totalCount:
        description: 'The amount total results matching the query'
        type: integer
        format: int64
        example: '887'
      currentPage:
        description: 'The current page number'
        type: integer
        format: int32
        example: '8'
      pageCount:
        description: 'The total number of pages'
        type: integer
        format: int32
        example: '8'
    required:
      - pageSize
      - count
      - totalCount
      - currentPage
      - pageCount
  SearchResult:
    type: object
    properties: 
      aggregations:
        description: 'The aggregations of the search result'
        type: array
        items:
          $ref: '#/definitions/AggregationResult'
      resultInfo:
        description: 'Search result specific information'
        $ref: '#/definitions/ResultInfo'
    required:
      - resultInfo
  RepositoryDescriptorSearchResult:
    allOf:
    - $ref: '#/definitions/SearchResult'
    properties:
      data:
        type: array
        items:
          $ref: '#/definitions/RepositoryDescriptor'
    required:
      - resultInfo
...
  RepositorySettings:
    type: object
    properties:
      publicAccess:
        description: 'True if the repository can be accessed publicly'
        type: boolean
        default: false
      prohibitsCommits:
        description: 'True if the repository is locked for further commits'
        type: boolean
        default: false
      commentApproval:
        description: 'True if the a comment needs to be approved before publication'
        type: boolean
        default: false
      maxSize:
        description: 'The maximum size in bytes if a limitation is set, 0 otherwise'
        type: integer
        format: int64
        default: 0
  RepositoryDescriptor:
    type: object
    properties:
      group: 
        description: 'The group of the repository'
        type: string
        example: 'examples'
      name: 
        description: 'The name of the repository'
        type: string
        example: 'reference_data'
      groupIsUserNamespace:
        description: 'True if the group is a user namespace'
        type: boolean
        default: false
      settings:
        $ref: '#/definitions/RepositorySettings'
    required:
      - group
      - name
...
parameters:
...
  page:
    name: 'page'
    description: 'page value for pagination, if the value is 0, no paging is applied and all values will be returned'
    in: query
    type: integer
    format: int32
    default: 1
  pageSize:
    name: 'pageSize'
    description: 'Defines the page size for pagination if page is not 0'
    in: query
    type: integer
    format: int32
    default: 10
  filter:
    name: 'filter'
    description: 'Filters results that contain the phrase in their name'
    in: query
    type: string
responses:
  NotLoggedIn:
    description: 'Not authenticated'
...
paths:
...
  /ws/repository:
    get:
      tags:
      - 'Repository'
      summary: 'List accessible repositories'
      description: 'Loads a list of repositories that are accessible for the client'
      operationId: 'getRepositories'
      produces:
      - application/json
      parameters:
      - $ref: '#/parameters/page'
      - $ref: '#/parameters/pageSize'
      - $ref: '#/parameters/filter'
      - name: 'group'
        description: 'Filters repositories by group (exact match)'
        in: query
        type: string
      - name: 'module'
        description: 'Module specific filter. If value is "REVIEW", the list of accessible repositories is filtered for those the user can manage tasks in'
        in: query
        type: string
        enum: ['REVIEW']
      responses:
        200:
          description: 'Returns a list of accessible repositories'
          schema:
            $ref: '#/definitions/RepositoryDescriptorSearchResult'
        401:
          $ref: '#/responses/NotLoggedIn'
...